Initial upload
My answer to a coder that was trying to get a small app to add two numbers together that were input by a user. Their code would loop, had the wrong way of dealing with the ValueError if the user input was not an integer or float. Didn't have a way to convert an int to float.
They include an exit option but it relied on the User press uppercase Q with no way of changing the user input to uppercase if the input were lowercase q. As it is more likely the input will be lower case, I set it that way.

The forum was ran by someone teaching python, via short video lessons, as I had already taken this course I knew what the coder was trying to attempt but went to complex without checking their code in small steps.