Indeed Practice Scrape
This was not a full fledged scrape of the site. As that could be frowned upon. 
It more to test that I could pull a specific job title and area, get the various details of a vacancy and present it in a nice manner.
Successful
To pull job title, company and salary
Unsuccessful
In obtaining the url link behind the title.
To be able to show more than three components at a time with the list comprehension